requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pyodbc==5.1.0
python-dotenv==1.0.1
